+++ CHANGELOG.md
@@ -14,10 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - perf: to_dict/dict_to_obj of OpenTriangleShell3D
 - Cylinder / Cone / HollowCylinder: from_center_point_and_axis
 - Cone: remove inheritance from RevolvedProfile
-- Ellipse2D: point_distance
+- Ellipse2D: point_distance, bounding rectangle, ellipse_intersections
 - Curve: local_discretization
+- Ellipse3D: line_intersections, linesegment_intersections, ellipse_intersections
 - BSplineSurface3D: partial removal of dependencies on geomdl objects
 - BSplineSurface3D: from_points_interpolation, from_points_approximation
+- 
 
 ### Fixed
 - Sweep with non smoth path
@@ -80,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Unittests
 - Cylinder / Cone / HollowCylinder
 - Ellipse2D: point_distance
+- Ellipse3D: test_ellipse_intersections, test_linesegment_intersections
 
 ## v0.12.0